ml_services/requirements.txt
pika
pillow
tensorflow
pandas
python-dotenv
keras-ocr
easyocr
torch 
torchvision 
torchaudio
sentence-transformers
numpy
